# packages/contract/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Arch-Systems Portal API",
    "version": "1.0.0",
    "description": "API for industrial operations portal - control room, drilling, engineering, safety, and production management"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT authentication token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/webhooks/{id}/logs": {
      "get": {
        "summary": "Get webhook delivery logs",
        "description": "Retrieve delivery logs for a specific webhook endpoint. Returns the last 50 log entries in reverse chronological order. Non-admins can only view logs for webhooks in their department.",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Webhook endpoint ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of delivery logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "webhook_endpoint_id": {
                            "type": "string"
                          },
                          "event_type": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "success",
                              "failed",
                              "retrying"
                            ]
                          },
                          "response_status": {
                            "type": "number",
                            "nullable": true
                          },
                          "error_message": {
                            "type": "string",
                            "nullable": true
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Webhook not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/telemetry/push": {
      "post": {
        "summary": "Push telemetry data to SCADA",
        "description": "Forward machine telemetry to FUXA SCADA server with two-level caching (in-memory + Redis). Accepts Supabase webhook payloads or direct tag updates. Deduplicates unchanged values.",
        "tags": [
          "Telemetry"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "description": "Supabase webhook payload (auto-detected)",
                    "required": [
                      "table",
                      "record"
                    ],
                    "properties": {
                      "table": {
                        "type": "string",
                        "enum": [
                          "machine_telemetry"
                        ]
                      },
                      "record": {
                        "type": "object",
                        "properties": {
                          "machine_id": {
                            "type": "string"
                          },
                          "engine_rpm": {
                            "type": "number"
                          },
                          "engine_temp": {
                            "type": "number"
                          },
                          "hydraulic_pressure": {
                            "type": "number"
                          },
                          "vibration_level": {
                            "type": "number"
                          },
                          "fuel_level": {
                            "type": "number"
                          },
                          "bit_depth": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  },
                  {
                    "type": "object",
                    "description": "Direct tag update",
                    "required": [
                      "name",
                      "value"
                    ],
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "Tag name (e.g., machine_123_engine_rpm)"
                      },
                      "value": {
                        "type": "number",
                        "description": "Tag value"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Telemetry processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "synced": {
                      "type": "boolean"
                    },
                    "cached": {
                      "type": "boolean",
                      "description": "True if value unchanged (deduplicated)"
                    },
                    "webhook": {
                      "type": "boolean",
                      "description": "True for webhook payload format"
                    },
                    "processed": {
                      "type": "integer",
                      "description": "Number of tags processed (webhook mode)"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tag": {
                            "type": "string"
                          },
                          "success": {
                            "type": "boolean"
                          },
                          "cached": {
                            "type": "boolean"
                          },
                          "error": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Internal server error or SCADA unreachable"
          }
        }
      }
    },
...
  },
...
}